pyproject.toml
[tool.poetry]
name = "data-spy"
version = "0.1.0"
description = ""
authors = ["Anthony Khoudary <ankhoudary12@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
snowflake-connector-python = {extras = ["pandas"], version = "^2.8.1"}
pandas = "^1.5.1"
python-decouple = "^3.6"
SQLAlchemy = "^1.4.43"
snowflake-sqlalchemy = "^1.4.3"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
black = "^22.10.0"
flake8-import-order = "^0.18.1"
flake8-black = "^0.3.3"
flake8-bugbear = "^22.10.27"
flake8-annotations = "^2.9.1"
flake8-docstrings = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
